+++ src/main/java/org/springframework/data/solr/core/query/AbstractQueryDecorator.java
@@ -150,7 +150,7 @@ public abstract class AbstractQueryDecorator implements Query {
 	}
 
 	@Override
-	public <T extends Query> T setOffset(Integer offset) {
+	public <T extends Query> T setOffset(Long offset) {
 		return this.query.setOffset(offset);
 	}
 
@@ -160,7 +160,7 @@ public abstract class AbstractQueryDecorator implements Query {
 	}
 
 	@Override
-	public Integer getOffset() {
+	public Long getOffset() {
 		return this.query.getOffset();
 	}
